backend/requirements.txt
anthropic==0.39.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain==0.3.7
langgraph==0.2.53
langchain-openai==0.2.6
pymupdf==1.24.10
python-docx==1.1.2
psycopg2-binary==2.9.9
pgvector==0.3.6
pydantic==2.9.2
requests==2.32.3
